+++ CHANGELOG.markdown
@@ -1,4 +1,4 @@
-# 1.0.1.pre
+# 1.0.1
 
 * Unicorn support (requires "preload_app true" in unicorn config file)
 * Fix for Thin detection - ensure it's actually running